Tower AI Logic
I have created a logic where a tower can shoot a projectile from a fire position and detect when an enemy is within range or view. This is done using a radius. This also allows me to use a Linerender component to show the radius which I will change later so It can be apart of the game.

The camera effects need to be changed with time but for now I am happy with the state of the game