It's Alive!!
Yes we now have AI. A very basic approach to them in the times we live in but an AI none the less.

What I have done here is used a Mix of 3 Elements to apply to an AI Unit and allow for Movement and Attack.

Each Unit will have the AI_Unit.cs attached to the Prefab. This sets up the AI, Allows for a shared trait (Movement and Death) and is a good way to have a basic yet shared behaviour.

Then we have the AI_Attack base class. I'm using polymorphism to get a behaviour that allows for each unit type to have a specific attack depending on what they are. This being said I can explain. A Heavy enemy like a Giant will be able to use Melee attack, but so can a Barbarian. Since I'm using a class called States in AI_Unit to set information like DMG, HP and TimeBetweenAttacks. I can also use the idea that each Unit will have different attacks without having to make a class for each enemy. If an archer can shoot and another troop can shoot then they can use the same Preset function and then have different art and state outputs.

Since this is now done I can allow for each tower to be destroyed and when this happens if my code is correct then the troops should move to the next tower and then we have some interactable Game Play.