TH Introduction 1
The TH the heart of the village is now working. By working I mean I moved around some already active functions that was making the camera behaviour change.

This is not the final changes to the TH since a lot of the functionality is clunky and needs a lot more tweaking and reworks