Camera Re Work 1
I have changed the Camera rotation functionality so now instead of having another object orbit around the desired Defence Piece that the player want's to upgrade the Camera now orbits independently while also having it's values reset when the player rather upgrades or leaves the upgrade menu.

The next piece of work that needs to be done is to make these camera transitions smoother right now they are slightly wild and loopy which I don't like.