Camera Re Work 2
The camera has been changed again and now the script structure is more modular. I have moved all the functionality over to the Upgrade helper which now controls the camera system behaviours. Using Enums to control when each function is called which allows for smoother transtions. I have also made a Blend asset which has proved to help support the transitions. More work will be done to these functions later but for now this works well for early development.

Next I need to set the Canon up to work with the script Made for Defences and then I can see any changes that need to be made for each Defence Object. Each defence tower have similar functions they shoot something so I will see if pre-sets need to be made which could be possible.