Animations
After passing my first mini milestone I decided to add some animations that i have gathered and changing the player character mesh to a simple dummy instead a cube. So far the movement kind of works but still needs tweaking and changes. I have a blend tree using a 1D blend that can blend from Idle walking and running and then go back to Idle when the input values are not being pressed.

I know eventually I need to blend these more naturally and decrease the value faster using a Lerp so the player does not instantly stop causing a form of jolt motion within the animations.

For now this works and next I need to try and add the combat system which will be interesting due to the amount of clips I plan to blend through