Cycle based Melee System W Anims
So this was hellish but also fun I made a cycle system that allows the player to Draw and Sheath their sword. But I wanted to play an inward and Outward attack back to back. Considering there are 18 animations for attack. I have managed to make a system that allows me to cycle through these animations randomly but within conditions that these animations do not get replayed until the cycle is over and also randomly picking each animation for different combos.

This works the only error is that if you spam attack you will make the player character look stupid. I am going to fix this later