Grid System and Melee Attack
In this push I have added a gird system that allows AI Defence units to snap onto the gird just like in clash of clans. I will eventually break all the functionality down so there is a cycle to this. I also added a very primitive Melee attack for now. Animations will need to come later for this to work.